Refactor: Move format_prop Static Method for Improved Reusability (#1714)
